requirements.txt
annotated-types==0.7.0
blinker==1.9.0
click==8.3.2
colorama==0.4.6
Flask==3.1.3
flask-cors==6.0.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
Werkzeug==3.1.8
anthropic==0.94.1
anyio==4.13.0
APScheduler==3.11.2
certifi==2026.2.25
charset-normalizer==3.4.7
distro==1.9.0
docstring_parser==0.18.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.14.0
playwright==1.58.0
pydantic==2.13.0
pydantic_core==2.46.0
pyee==13.0.1
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
sniffio==1.3.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
tzlocal==5.3.1
urllib3==2.6.3
